Revert "sum3 example and pytest"
