--- .vscode/settings.json
@@ -1,14 +1,0 @@
-{
-    "python.analysis.extraPaths": [
-        "./src",
-        "./src/arrays",
-    ],
-    "python.testing.pytestArgs": [
-        ".",
-        
-    ],
-    "python.linting.enabled": true,
-    "python.testing.unittestEnabled": true,
-    "python.testing.pytestEnabled": true,
-    // "python.testing.cwd": "${workspaceFolder}/tests",
-}
