+++ arrays/container_With_Most_Water_11.py
@@ -27,7 +27,6 @@ from typing import List
 # plt.imshow(cv2.cvtColor(im_resized, cv2.COLOR_BGR2BGRA))
 
 class Water_Containers():
-# class Water_Containers():
     def maxArea(self, height: List[int]) -> int:
         n =len(height) - 1
         p1, p2 = 0, n
